use natchez-noop until we can figure out XRay tracing again
Our XRay Natchez library was only published to Bintray, and I think it
had thread safety issues, so it might not be worth republishing to Maven
Central. There's another XRay Natchez implementation that might work,
but this at least gets the project building so we can move it to
GitHub Actions.